+++ src/models/basic_summary_entry.rs
@@ -26,7 +26,7 @@ pub struct BasicSummaryEntry {
     pub monetary_value: Option<f64>,
     /// The currency ID of the associated currency.
     #[serde(rename = "currency_id", skip_serializing_if = "Option::is_none")]
-    pub currency_id: Option<f32>,
+    pub currency_id: Option<i32>,
     #[serde(rename = "currency_code", skip_serializing_if = "Option::is_none")]
     pub currency_code: Option<String>,
     #[serde(rename = "currency_symbol", skip_serializing_if = "Option::is_none")]
